auth_user_service/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2==2.9.6, <3.0.0
drf-spectacular==0.26.1, <0.27.0
djangorestframework-simplejwt==5.5.0, <6.0.0
jwt==3.3.0, <4.0.0